Robot config changes per code review - removed default robot actuator pins, etc.
Removed default robot actuator pins.
Flattened checksum array.
Optimized checksum array for k_max_actuators==3.
Added compile time static_assert if checksum array is smaller than
k_max_actuators.
Removed printf warning that arm solution wants more than
k_max_actuators.

+++ src/modules/robot/ActuatorCoordinates.h
@@ -15,7 +15,7 @@
 
 //The subset in use is determined by the arm solution's get_actuator_count().
 //Keep MAX_ROBOT_ACTUATORS as small as practical it impacts block size and therefore free memory.
-const int k_max_actuators = MAX_ROBOT_ACTUATORS;
+const size_t k_max_actuators = MAX_ROBOT_ACTUATORS;
 typedef struct std::array<float, k_max_actuators> ActuatorCoordinates;
 
 #endif
